wip: refactor of flux-go
This refactors flux-go so it better can provide two use cases:

flux as a service "FaaS" (ha), meaning a grpc server with access
to a flux handle. And then go bindings / sdk to core flux operations.
Both assume that there is a flux handle somewhere we want to interact
with. For the latter, I think the easiest thing will generally
be to use rpc pack and unpack, written as c functions in the go code
(via cgo). The reason is because the functions are variadic and thus
could not be called directly from cgo. If we try to use functions
directly that return a flux future, to "get" that we need to handle
some weird const void **result and I just have no idea.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>